requirements.txt
websocket-client==1.9.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==1.5.3